+++ pontos/version/__main__.py
@@ -19,12 +19,21 @@
 
 import sys
 
+from pathlib import Path
 from .version import PontosVersionCommand
+from .cmake_version import CMakeVersionCommand
 
 
 def main():
-    cmd = PontosVersionCommand()
-    sys.exit(cmd.run())
+    available_cmds = [
+        ('CMakeLists.txt', CMakeVersionCommand),
+        ('pyproject.toml', PontosVersionCommand),
+    ]
+    for fileName, cmd in available_cmds:
+        project_definition_path = Path.cwd() / fileName
+        if project_definition_path.exists():
+            sys.exit(cmd().run())
+    sys.exit("No command found")
 
 
 if __name__ == '__main__':